# dbsync-api/swagger.yaml
openapi: 3.0.0
info:
  title: Stake Address API
  version: 1.0.0
  description: API for retrieving stake registration and deregistration information by address.

paths:
  /api/stake-address:
    get:
      summary: Retrieve stake address information
      description: Get the latest stake registration and deregistration for a given stake address.
      parameters:
        - name: address
          in: query
          required: true
          description: The stake address in hexadecimal or Bech32 format.
          schema:
            type: string
      responses:
        '200':
          description: Successful response with stake address data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration:
                    type: object
                    nullable: true
                    properties:
                      epoch_no:
                        type: integer
                        description: Epoch number.
                      slot_no:
                        type: integer
                        description: Slot number.
                      block_no:
                        type: integer
                        description: Slot Number.
                      time:
                        type: string
                        description: Time of transaction.
                      tx_hash:
                        type: string
                        description: Transaction hash.
                  deregistration:
                    type: object
                    nullable: true
                    properties:
                      epoch_no:
                        type: integer
                        description: Epoch number.
                      slot_no:
                        type: integer
                        description: Slot number.
                      block_no:
                        type: integer
                        description: Slot Number.
                      time:
                        type: string
                        description: Time of transaction.
                      tx_hash:
                        type: string
                        description: Transaction hash.
        '400':
          description: Bad Request. The request is missing required parameters or has invalid data.
        '404':
          description: Not Found. The stake address does not exist in the database.
        '500':
          description: Internal Server Error. An unexpected error occurred while processing the request.

  /api/delegation:
    get:
      summary: Get latest delegation information
      description: Fetches the latest delegation and pool delegation details for the given address.
      operationId: getDelegation
      parameters:
        - name: address
          in: query
          required: true
          description: Bech32 or hex-encoded address to query.
          schema:
            type: string
            example: ""
      responses:
        '200':
          description: Successful response with delegation and pool information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  drep:
                    type: array
                    items:
                      type: object
                      properties:
                        drep_id:
                          type: string
                          description: The view of the DRep.
                        epoch_no:
                          type: integer
                          description: The epoch number of the delegation.
                        slot_no:
                          type: integer
                          description: The slot number of the delegation.
                        block_no:
                          type: integer
                          description: The block number.
                        time:
                          type: string
                          format: date-time
                          description: The time of the block.
                        tx_hash:
                          type: string
                          description: The transaction hash.
                  pool:
                    type: array
                    items:
                      type: object
                      properties:
                        pool_id:
                          type: string
                          description: The view of the pool.
                        epoch_no:
                          type: integer
                          description: The epoch number of the delegation.
                        slot_no:
                          type: integer
                          description: The slot number.
                        block_no:
                          type: integer
                          description: The block number.
                        time:
                          type: string
                          format: date-time
                          description: The time of the block.
                        tx_hash:
                          type: string
                          description: The transaction hash.

        '400':
          description: Bad request. The `address` query parameter is missing or invalid.
        '500':
          description: Internal server error.

  /api/drep/{id}:
    get:
      summary: Get DRep details
      description: Fetches DRep details including registration status, metadata, voting power, and transaction hashes for the provided address.
      operationId: getDrepDetails
      parameters:
        - name: id
          in: path
          required: true
          description: Hex-encoded address to query DRep details.
          schema:
            type: string
            example: ""
      responses:
        '200':
          description: Successful response with DRep and SoleVoter details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  isRegisteredAsDRep:
                    type: boolean
                    description: Whether the address is currently registered as a DRep.
                  wasRegisteredAsDRep:
                    type: boolean
                    description: Whether the address was previously registered as a DRep.
                  isRegisteredAsSoleVoter:
                    type: boolean
                    description: Whether the address is currently registered as a Sole Voter.
                  wasRegisteredAsSoleVoter:
                    type: boolean
                    description: Whether the address was previously registered as a Sole Voter.
                  deposit:
                    type: integer
                    description: The current deposit value.
                  url:
                    type: string
                    description: The metadata URL.
                  dataHash:
                    type: string
                    description: Hash of the metadata.
                  votingPower:
                    type: integer
                    description: The current voting power.
                  dRepRegisterTxHash:
                    type: string
                    description: Transaction hash for DRep registration.
                  dRepRetireTxHash:
                    type: string
                    description: Transaction hash for DRep retirement.
                  soleVoterRegisterTxHash:
                    type: string
                    description: Transaction hash for Sole Voter registration.
                  soleVoterRetireTxHash:
                    type: string
                    description: Transaction hash for Sole Voter retirement.

        '400':
          description: Bad Request. Address is invalid or missing.
        '500':
          description: Internal Server Error. Something went wrong with the query or database.

  /api/address/balance:
    get:
      summary: Get faucet balance for a given address
      description: Retrieves the balance for a specified stake address.
      parameters:
        - in: query
          name: address
          schema:
            type: string
          required: true
          description: The stake address in hex format to query the balance.
      responses:
        '200':
          description: The faucet balance for the given address
          content:
            application/json:
              schema:
                type: integer
                description: The balance in lovelace (smallest unit of ADA)
                example: 1000000000
        '400':
          description: Invalid address provided
        '500':
          description: Internal server error
  /api/drep:
    get:
      summary: Fetch DRep list
      description: Fetches the drep list.
      parameters:
        - in: query
          name: page
          schema:
            type: number
            default: 1
          required: false
        - in: query
          name: sort
          schema:
            type: string
            enum: [ VotingPower, RegistrationDate ]
        - in: query
          name: status
          schema:
            type: string
            enum: [ Active, Inactive, Retired ]
        - in: query
          name: size
          schema:
            type: number
            default: 10
          required: false
        - in: query
          name: dRepId
          description: Search by dRepId
          schema:
            type: string
            default: ''
          required: false
      responses:
        200:
          description: "A paginated list of DRep details."
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  items:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                          drepId:
                            type: "string"
                            example: "string"
                            description: "DRep ID in hexadecimal format."
                          view:
                            type: "string"
                            example: "string"
                            description: "DRep's view identifier."
                          url:
                            type: "string"
                            format: "uri"
                            example: "string"
                            description: "URL to the DRep's metadata."
                          metadataHash:
                            type: "string"
                            example: "string"
                            description: "Hash of the metadata."
                          deposit:
                            type: "integer"
                            format: "int64"
                            example: 0
                            description: "Deposit amount."
                          votingPower:
                            type: "integer"
                            format: "int64"
                            example: 0
                            description: "DRep's amount."
                          active:
                            type: "boolean"
                            example: false
                            description: "Indicates if the DRep is active."
                          latestTxHash:
                            type: "string"
                            example: "string"
                            description: "Hash of the latest transaction."
                          latestRegistrationDate:
                            type: "string"
                            format: "date-time"
                            example: "0000-00-00T00:00:00"
                            description: "Date and time of the latest registration."
                  page:
                    type: "integer"
                    example: 0
                    description: "Current page number."
                  size:
                    type: "integer"
                    example: 0
                    description: "Number of items returned per page."
                  total:
                    type: "integer"
                    example: 0
                    description: "Total number of items"
        '400':
          description: Invalid address provided
        '500':
          description: Internal server error
  /api/proposal:
    get:
      summary: Fetch Proposal list
      description: Fetches the Proposal list.
      parameters:
        - in: query
          name: page
          schema:
            type: number
            default: 1
          required: false
        - in: query
          name: size
          schema:
            type: number
            default: 10
          required: false
        - in: query
          name: sort
          schema:
            type: string
            enum: [ CreatedDate, ExpiryDate ]
        - in: query
          name: type
          schema:
            type: string
            enum: [ParameterChange, HardForkInitiation, TreasuryWithdrawals ,NoConfidence ,NewCommittee ,NewConstitution ,InfoAction]
        - in: query
          name: proposal
          description: Search by proposal id
          schema:
            type: string
            default: ''
          required: false
      responses:
        200:
          description: "A paginated list of Proposal details."
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  items:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: integer
                          example: 418
                        txHash:
                          type: string
                          example: ""
                        index:
                          type: integer
                          example: 0
                        type:
                          type: string
                          example: ""
                        details:
                          type: object
                          additionalProperties: true
                          example: { }
                        expiryDate:
                          type: string
                          format: date-time
                          example: "2024-10-08T07:25:57.226978"
                        expiryEpochNon:
                          type: integer
                          example: 481
                        createdDate:
                          type: string
                          format: date-time
                          example: "2024-09-23T10:35:09"
                        createdEpochNo:
                          type: integer
                          example: 466
                        url:
                          type: string
                          example: ""
                        metadataHash:
                          type: string
                          example: "1111111111111111111111111111111111111111111111111111111111111111"
                        protocolParams:
                          type: object
                          nullable: true
                        title:
                          type: string
                          nullable: true
                        abstract:
                          type: string
                          nullable: true
                        motivation:
                          type: string
                          nullable: true
                        rationale:
                          type: string
                          nullable: true
                        dRepYesVotes:
                          type: integer
                          example: 0
                        dRepNoVotes:
                          type: integer
                          example: ''
                        dRepAbstainVotes:
                          type: integer
                          example: ''
                        poolYesVotes:
                          type: integer
                          example: 0
                        poolNoVotes:
                          type: integer
                          example: 0
                        poolAbstainVotes:
                          type: integer
                          example: 0
                        ccYesVotes:
                          type: integer
                          example: 0
                        ccNoVotes:
                          type: integer
                          example: 0
                        ccAbstainVotes:
                          type: integer
                          example: 0
                        prevGovActionIndex:
                          type: integer
                          nullable: true
                        prevGovActionTxHash:
                          type: string
                          nullable: true
                  page:
                    type: "integer"
                    example: 0
                    description: "Current page number."
                  size:
                    type: "integer"
                    example: 0
                    description: "Number of items returned per page."
                  total:
                    type: "integer"
                    example: 0
                    description: "Total number of items"
        '400':
          description: Invalid address provided
        '500':
          description: Internal server error